+++ metatensor-torch/CHANGELOG.md
@@ -14,12 +14,14 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 
 ### Changed
 
-- `MetatensorAtomisticModel.save()` always saves models on the CPU.
-
 ### Deprecated
 
 ### Removed
 -->
+
+### Changed
+
+- `MetatensorAtomisticModel.save()` always saves models on the CPU.
 
 ## [Version 0.5.2](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-torch-v0.5.2) - 2024-06-21