+++ metatensor-torch/CHANGELOG.md
@@ -19,6 +19,10 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 ### Removed
 -->
 
+### Changed
+
+- `MetatensorAtomisticModel.save()` always saves models on the CPU.
+
 ## [Version 0.5.2](https://github.com/lab-cosmo/metatensor/releases/tag/metatensor-torch-v0.5.2) - 2024-06-21
 
 ### Added